+++ R/wfs_ad.R
@@ -8,7 +8,8 @@
 #' @seealso [sf::st_bbox()]
 #'
 #' @return A \CRANpkg{sf} object.
-#' @source [SITNA – Catastro de Navarra](https://sitna.navarra.es/geoportal/)
+#' @source
+#' [SITNA – Catastro de Navarra](https://geoportal.navarra.es/es/inspire)
 #'
 #' @details
 #'